requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Manipulation and Numerical Computing
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.5.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.2.0

# Progress bars and helpers
tqdm>=4.65.0